Move protbuf-specific code to o.p.c.common.types.pb
