+++ core/src/main/java/org/projectnessie/cel/common/types/ref/TypeAdapterSupport.java
@@ -13,14 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.cel.common.types;
+package org.projectnessie.cel.common.types.ref;
 
 import static org.projectnessie.cel.common.types.BytesT.bytesOf;
 import static org.projectnessie.cel.common.types.DoubleT.doubleOf;
 import static org.projectnessie.cel.common.types.DurationT.durationOf;
-import static org.projectnessie.cel.common.types.Err.anyWithEmptyType;
-import static org.projectnessie.cel.common.types.Err.newErr;
-import static org.projectnessie.cel.common.types.Err.unknownType;
 import static org.projectnessie.cel.common.types.IntT.intOf;
 import static org.projectnessie.cel.common.types.ListT.newGenericArrayList;
 import static org.projectnessie.cel.common.types.ListT.newJSONList;
@@ -33,14 +30,11 @@ import static org.projectnessie.cel.common.types.TimestampT.ZoneIdZ;
 import static org.projectnessie.cel.common.types.TimestampT.timestampOf;
 import static org.projectnessie.cel.common.types.Types.boolOf;
 import static org.projectnessie.cel.common.types.UintT.uintOf;
-import static org.projectnessie.cel.common.types.pb.TypeDescription.typeNameFromMessage;
 
-import com.google.api.expr.v1alpha1.Value;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.Descriptors.EnumValueDescriptor;
 import com.google.protobuf.EnumValue;
 import com.google.protobuf.ListValue;
-import com.google.protobuf.Message;
 import com.google.protobuf.NullValue;
 import com.google.protobuf.Struct;
 import com.google.protobuf.UInt32Value;
@@ -56,31 +50,16 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.BiFunction;
 import org.projectnessie.cel.common.ULong;
-import org.projectnessie.cel.common.types.pb.Db;
-import org.projectnessie.cel.common.types.pb.TypeDescription;
-import org.projectnessie.cel.common.types.ref.TypeAdapter;
-import org.projectnessie.cel.common.types.ref.Val;
+import org.projectnessie.cel.common.types.DoubleT;
+import org.projectnessie.cel.common.types.IntT;
+import org.projectnessie.cel.common.types.NullT;
+import org.projectnessie.cel.common.types.pb.DefaultTypeAdapter;
 
-/** defaultTypeAdapter converts go native types to CEL values. */
-public class DefaultTypeAdapter implements TypeAdapter {
-  /** DefaultTypeAdapter adapts canonical CEL types from their equivalent Go values. */
-  public static final DefaultTypeAdapter Instance = new DefaultTypeAdapter(Db.defaultDb);
-
-  private final Db db;
-
-  private DefaultTypeAdapter(Db db) {
-    this.db = db;
-  }
-
-  /** NativeToValue implements the ref.TypeAdapter interface. */
-  @Override
-  public Val nativeToValue(Object value) {
-    Val val = nativeToValue(db, this, value);
-    if (val != null) {
-      return val;
-    }
-    return Err.unsupportedRefValConversionErr(value);
-  }
+/**
+ * Helper class for {@link TypeAdapter} implementations to convert from a Java type to a CEL type.
+ */
+public final class TypeAdapterSupport {
+  private TypeAdapterSupport() {}
 
   private static final Map<Class<?>, BiFunction<TypeAdapter, Object, Val>> NativeToValueExact =
       new IdentityHashMap<>();
@@ -138,40 +117,7 @@ public class DefaultTypeAdapter implements TypeAdapter {
         });
   }
 
-  /**
-   * nativeToValue returns the converted (ref.Val, true) of a conversion is found, otherwise (nil,
-   * false)
-   */
-  public static Val nativeToValue(Db db, TypeAdapter a, Object value) {
-    Val v = maybeNativeToValue(a, value);
-    if (v != null) {
-      return v;
-    }
-
-    // additional specializations may be added upon request / need.
-    if (value instanceof Val) {
-      return (Val) value;
-    }
-    if (value instanceof Message) {
-      Message msg = (Message) value;
-      String typeName = typeNameFromMessage(msg);
-      if (typeName.isEmpty()) {
-        return anyWithEmptyType();
-      }
-      TypeDescription type = db.describeType(typeName);
-      if (type == null) {
-        return unknownType(typeName);
-      }
-      value = type.maybeUnwrap(db, msg);
-      if (value instanceof Message) {
-        value = type.maybeUnwrap(db, (Message) value);
-      }
-      return a.nativeToValue(value);
-    }
-    return newErr("unsupported conversion from '%s' to value", value.getClass());
-  }
-
-  static Val maybeNativeToValue(TypeAdapter a, Object value) {
+  public static Val maybeNativeToValue(TypeAdapter a, Object value) {
     if (value == null) {
       return NullT.NullValue;
     }
@@ -209,37 +155,5 @@ public class DefaultTypeAdapter implements TypeAdapter {
     }
 
     return null;
-  }
-
-  static Object maybeUnwrapValue(Object value) {
-    if (value instanceof Value) {
-      Value v = (Value) value;
-      switch (v.getKindCase()) {
-        case BOOL_VALUE:
-          return v.getBoolValue();
-        case BYTES_VALUE:
-          return v.getBytesValue();
-        case DOUBLE_VALUE:
-          return v.getDoubleValue();
-        case INT64_VALUE:
-          return v.getInt64Value();
-        case LIST_VALUE:
-          return v.getListValue();
-        case NULL_VALUE:
-          return v.getNullValue();
-        case MAP_VALUE:
-          return v.getMapValue();
-        case STRING_VALUE:
-          return v.getStringValue();
-        case TYPE_VALUE:
-          return Types.getTypeByName(v.getTypeValue());
-        case UINT64_VALUE:
-          return ULong.valueOf(v.getUint64Value());
-        case OBJECT_VALUE:
-          return v.getObjectValue();
-      }
-    }
-
-    return value;
   }
 }
