+++ core/src/main/java/org/projectnessie/cel/common/types/ref/TypeDescription.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2021 The Authors of CEL-Java
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.cel.common.types.ref;
+
+/**
+ * TypeDescription is a collection of type metadata relevant to expression checking and evaluation.
+ */
+public interface TypeDescription {
+
+  /** Name returns the fully-qualified name of the type. */
+  String name();
+
+  /** ReflectType returns the Golang reflect.Type for this type. */
+  Class<?> reflectType();
+}
