+++ core/src/test/java/org/projectnessie/cel/CELTest.java
@@ -54,8 +54,8 @@ import static org.projectnessie.cel.common.types.Err.isError;
 import static org.projectnessie.cel.common.types.Err.newErr;
 import static org.projectnessie.cel.common.types.Err.valOrErr;
 import static org.projectnessie.cel.common.types.IntT.IntOne;
-import static org.projectnessie.cel.common.types.ProtoTypeRegistry.newEmptyRegistry;
 import static org.projectnessie.cel.common.types.StringT.stringOf;
+import static org.projectnessie.cel.common.types.pb.ProtoTypeRegistry.newEmptyRegistry;
 import static org.projectnessie.cel.interpreter.Activation.emptyActivation;
 import static org.projectnessie.cel.interpreter.Interpretable.newConstValue;
 import static org.projectnessie.cel.parser.Macro.newReceiverMacro;
@@ -80,11 +80,11 @@ import org.projectnessie.cel.Program.EvalResult;
 import org.projectnessie.cel.checker.Decls;
 import org.projectnessie.cel.common.operators.Operator;
 import org.projectnessie.cel.common.types.BoolT;
-import org.projectnessie.cel.common.types.DefaultTypeAdapter;
 import org.projectnessie.cel.common.types.IntT;
 import org.projectnessie.cel.common.types.Overloads;
 import org.projectnessie.cel.common.types.StringT;
 import org.projectnessie.cel.common.types.UnknownT;
+import org.projectnessie.cel.common.types.pb.DefaultTypeAdapter;
 import org.projectnessie.cel.common.types.ref.TypeRegistry;
 import org.projectnessie.cel.common.types.ref.Val;
 import org.projectnessie.cel.common.types.traits.Container;
