+++ core/src/test/java/org/projectnessie/cel/interpreter/AttributePatternsTest.java
@@ -18,8 +18,8 @@ package org.projectnessie.cel.interpreter;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.projectnessie.cel.Util.mapOf;
-import static org.projectnessie.cel.common.types.ProtoTypeRegistry.newRegistry;
 import static org.projectnessie.cel.common.types.UnknownT.unknownOf;
+import static org.projectnessie.cel.common.types.pb.ProtoTypeRegistry.newRegistry;
 import static org.projectnessie.cel.interpreter.Activation.emptyActivation;
 import static org.projectnessie.cel.interpreter.Activation.newPartialActivation;
 import static org.projectnessie.cel.interpreter.AttributePattern.newAttributePattern;
