+++ core/src/test/java/org/projectnessie/cel/interpreter/AttributesTest.java
@@ -26,9 +26,9 @@ import static org.projectnessie.cel.common.types.BoolT.True;
 import static org.projectnessie.cel.common.types.Err.newErr;
 import static org.projectnessie.cel.common.types.IntT.IntType;
 import static org.projectnessie.cel.common.types.IntT.intOf;
-import static org.projectnessie.cel.common.types.ProtoTypeRegistry.newRegistry;
 import static org.projectnessie.cel.common.types.StringT.stringOf;
 import static org.projectnessie.cel.common.types.UnknownT.unknownOf;
+import static org.projectnessie.cel.common.types.pb.ProtoTypeRegistry.newRegistry;
 import static org.projectnessie.cel.interpreter.Activation.emptyActivation;
 import static org.projectnessie.cel.interpreter.Activation.newActivation;
 import static org.projectnessie.cel.interpreter.Activation.newPartialActivation;
@@ -60,8 +60,8 @@ import org.projectnessie.cel.checker.CheckerEnv;
 import org.projectnessie.cel.checker.Decls;
 import org.projectnessie.cel.common.Source;
 import org.projectnessie.cel.common.containers.Container;
-import org.projectnessie.cel.common.types.DefaultTypeAdapter;
 import org.projectnessie.cel.common.types.UnknownT;
+import org.projectnessie.cel.common.types.pb.DefaultTypeAdapter;
 import org.projectnessie.cel.common.types.ref.TypeRegistry;
 import org.projectnessie.cel.common.types.ref.Val;
 import org.projectnessie.cel.interpreter.AttributeFactory.Attribute;
