+++ core/src/test/java/org/projectnessie/cel/interpreter/InterpreterTest.java
@@ -35,12 +35,12 @@ import static org.projectnessie.cel.common.types.IntT.IntZero;
 import static org.projectnessie.cel.common.types.IntT.intOf;
 import static org.projectnessie.cel.common.types.NullT.NullType;
 import static org.projectnessie.cel.common.types.NullT.NullValue;
-import static org.projectnessie.cel.common.types.ProtoTypeRegistry.newEmptyRegistry;
-import static org.projectnessie.cel.common.types.ProtoTypeRegistry.newRegistry;
 import static org.projectnessie.cel.common.types.StringT.StringType;
 import static org.projectnessie.cel.common.types.StringT.stringOf;
 import static org.projectnessie.cel.common.types.Types.boolOf;
 import static org.projectnessie.cel.common.types.UintT.uintOf;
+import static org.projectnessie.cel.common.types.pb.ProtoTypeRegistry.newEmptyRegistry;
+import static org.projectnessie.cel.common.types.pb.ProtoTypeRegistry.newRegistry;
 import static org.projectnessie.cel.common.types.traits.Trait.AdderType;
 import static org.projectnessie.cel.common.types.traits.Trait.NegatorType;
 import static org.projectnessie.cel.interpreter.Activation.emptyActivation;
@@ -94,13 +94,13 @@ import org.projectnessie.cel.checker.Decls;
 import org.projectnessie.cel.common.Source;
 import org.projectnessie.cel.common.ULong;
 import org.projectnessie.cel.common.containers.Container;
-import org.projectnessie.cel.common.types.DefaultTypeAdapter;
 import org.projectnessie.cel.common.types.DurationT;
 import org.projectnessie.cel.common.types.Err;
 import org.projectnessie.cel.common.types.ListT;
 import org.projectnessie.cel.common.types.StringT;
 import org.projectnessie.cel.common.types.TimestampT;
 import org.projectnessie.cel.common.types.UnknownT;
+import org.projectnessie.cel.common.types.pb.DefaultTypeAdapter;
 import org.projectnessie.cel.common.types.ref.TypeRegistry;
 import org.projectnessie.cel.common.types.ref.Val;
 import org.projectnessie.cel.common.types.traits.Adder;
